requirements-production.txt
# Production-specific requirements
# Run: pip install -r requirements-production.txt

# Web server
gunicorn==22.0.0
whitenoise==6.6.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Caching & Redis
django-redis==5.4.0
redis==5.0.1

# Celery (already in requirements.txt)
celery==5.4.0
django-celery-beat==2.5.0

# Email (optional)
sendgrid==6.11.0

# Error tracking (optional)
sentry-sdk==1.45.2

# Security
django-cors-headers==4.3.1
django-ratelimit==4.1.0

# AWS S3 (optional)
boto3==1.34.67
django-storages==1.14.2

# Monitoring
django-extensions==3.2.3

# All base requirements
-r requirements.txt
